Update NDK Callback. Reorg NDK files.
