--- Assets/Inworld/Inworld.NDK/Plugins/ThirdParty.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f68d8a3ac6123c44ba8011162958013d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
