+++ Assets/Inworld/Inworld.NDK/Scenes/NDKSampleBasic.unity
@@ -240,7 +240,7 @@ MonoBehaviour:
   m_Client: {fileID: 359806128}
   m_AudioCapture: {fileID: 0}
   m_CharacterHandler: {fileID: 359806131}
-  m_GameData: {fileID: 0}
+  m_GameData: {fileID: 11400000, guid: ddfc6d5fd8aef7e4d96034cee968d78a, type: 2}
   m_SceneFullName: workspaces/unitytest/scenes/unity_gallery
   m_AutoStart: 1
 --- !u!114 &359806128
