--- Assets/Inworld/Inworld.NDK/Scripts/NDK/InworldNDK.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2b8e03a4fdc643ca80d98eeb53487065
-timeCreated: 1694497241
