+++ Assets/Inworld/Inworld.NDK/Scripts/NDK/ProcessingAudioChunk.cs
@@ -1,0 +1,38 @@
+namespace Inworld.NDK
+{
+    public class ProcessingAudioChunk
+    {
+        readonly Inworld.Packet.AudioPacket m_CurrentAudioPacket;
+        readonly int m_PhonemeCount;
+        readonly bool m_Initialized;
+
+        public ProcessingAudioChunk()
+        {
+            
+        }
+        public ProcessingAudioChunk(NDKPacket rhs)
+        {
+            m_CurrentAudioPacket = InworldNDK.From.NDKAudioChunk(rhs);
+            m_PhonemeCount = rhs.audioPacket.phonemeCount;
+            m_Initialized = true;
+        }
+        public void ReceivePhoneme(PhonemeInfo phonemeInfo)
+        {
+            if (phonemeInfo.packetID != m_CurrentAudioPacket.packetId.packetId)
+                return;
+            m_CurrentAudioPacket.dataChunk.additionalPhonemeInfo.Add(new Inworld.Packet.PhonemeInfo
+            {
+                phoneme = phonemeInfo.code,
+                startOffset = phonemeInfo.timeStamp
+            });
+            if (m_CurrentAudioPacket.dataChunk.additionalPhonemeInfo.Count != m_PhonemeCount)
+                return;
+            ToInworldPacket();
+        }
+        public void ToInworldPacket()
+        {
+            if (m_Initialized && InworldController.Client is InworldNDKClient ndkClient)
+                ndkClient.Enqueue(m_CurrentAudioPacket);
+        }
+    }
+}
